Implement Row Reader
Row interface provides an access to the entire row.
Each column of the row can be accessed multiple times.
Providing a wrong column name to Row functions will cause a panic.

(*Rows).Row() function can be executed multiple times
and can be used together with column specific functions of (*Rows)

Changes to (*Rows).Next()
Previous implementation stops the loop if there is a parsing error.
Now the loop is canceled only when there is a reading error from the stream.
However, parsing error will be still returned from (*Rows).LastError()
The reason is that function that caused a parsing error can be replaced with another one.
For instance:
  row := rows.Row()
  age := row.Int("age")
  if row.LastError() != nil {
      ageStr := row.String("age")
      // custom convertor
  }